backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==15.0.1
aiosqlite==0.21.0
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
pyotp==2.9.0
python-dotenv==1.1.0
pydantic==2.11.1
pydantic-settings==2.8.1
docker==7.1.0
